requirements.txt

solana
solders
python-dotenv
colorama
